Updated how development server starts. Initial setup of multiple game-workers running in a multi-core processor server.

+++ scripts/dev.js
@@ -1,8 +1,8 @@
 // npm
 import { EventEmitter } from 'events';
-import ParcelBundler from 'parcel-bundler';
 import path from 'path';
 import cp from 'child_process';
+import fs from 'fs';
 import treeKill from 'tree-kill';
 
 // dev function
@@ -11,41 +11,51 @@ const dev = async () => {
 	const events = new EventEmitter();
 
 	// bundlers
-	const server = new ParcelBundler(path.resolve(__dirname, '..', 'src', 'server', 'index.ts'), {
-		outDir: path.resolve(__dirname, '..', 'dist'),
-		outFile: 'server.js',
-		watch: true,
-		cache: false,
-		target: 'node',
-		bundleNodeModules: true,
-		hmr: false,
-		logLevel: 0,
-		autoInstall: false,
-		throwErrors: false,
+	const server = cp.exec('yarn dev-server');
+
+	// server output
+	server.stdout.on('data', (data) => {
+		if (typeof data !== 'string') {
+			console.log('[server_build]:', data);
+		} else {
+			data
+				.trim()
+				.split('\n')
+				.forEach((data) => {
+					console.log('[server_build]:', data);
+
+					if (data.indexOf('Built in ') !== -1) {
+						events.emit('server-bundle-end');
+					}
+				});
+		}
 	});
-
-	// bundlers events
-	server.on('buildEnd', () => events.emit('server-bundle-end'));
-
-	// run bundlers
-	await server.bundle();
 
 	// run server
 	const runServer = () => {
-		// start process
-		const { pid } = cp.spawn('node', [path.resolve(__dirname, '..', 'dist', 'server.js')], { stdio: 'inherit', env: { NODE_ENV: 'development' } });
+		// server path
+		const serverPath = path.resolve(__dirname, '..', 'dist', 'server.js');
 
-		// in server build
-		events.once('server-bundle-end', () => {
-			// log
-			console.log('server restarting...');
+		// check if server was built
+		if (fs.existsSync(serverPath)) {
+			// start process
+			const { pid } = cp.spawn('node', [serverPath], { stdio: 'inherit', env: { NODE_ENV: 'development' } });
 
-			// kill process
-			treeKill(pid, () => {
-				// run server runner
-				runServer();
+			// in server build
+			events.once('server-bundle-end', () => {
+				// log
+				console.log('server restarting...');
+
+				// kill process
+				treeKill(pid, () => {
+					// run server runner
+					runServer();
+				});
 			});
-		});
+		} else {
+			// run server runner
+			runServer();
+		}
 	};
 
 	// run server runner
